% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Title}
\usage{
hello(msg)
}
\arguments{
\item{msg}{Input that should be printed in green.}
}
\value{
Prints your message in green.
}
\description{
Title
}
\examples{
hello("hello")
}
