% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_fun.R
\name{new_function}
\alias{new_function}
\title{Title}
\usage{
new_function(msg)
}
\arguments{
\item{msg}{Input that should be printed in magenta.}
}
\value{
Prints your message in magenta
}
\description{
Title
}
\examples{
hello("hello")
}
